backend/requirements.txt
openai==1.6.1
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
python-dotenv==1.0.0